#6 Updated tests to catch messaging file bugs. Moved messaging to pkg dir.
