+++ internal/app/dispatcher/providers/args.go
@@ -3,8 +3,8 @@ package providers
 import (
 	"github.com/Azure/azure-sdk-for-go/services/servicebus/mgmt/2017-04-01/servicebus"
 	"github.com/joho/godotenv"
-	"github.com/lawrencegripper/ion/internal/app/dispatcher/messaging"
 	"github.com/lawrencegripper/ion/internal/pkg/common"
+	"github.com/lawrencegripper/ion/internal/pkg/messaging"
 	"github.com/lawrencegripper/ion/internal/pkg/types"
 	log "github.com/sirupsen/logrus"
