+++ internal/app/dispatcher/providers/azurebatch.go
@@ -12,7 +12,7 @@ import (
 	"github.com/Azure/go-autorest/autorest"
 	"github.com/Azure/go-autorest/autorest/to"
 	"github.com/lawrencegripper/ion/internal/app/dispatcher/helpers"
-	"github.com/lawrencegripper/ion/internal/app/dispatcher/messaging"
+	"github.com/lawrencegripper/ion/internal/pkg/messaging"
 	"github.com/lawrencegripper/ion/internal/pkg/types"
 	"github.com/lawrencegripper/pod2docker"
 	log "github.com/sirupsen/logrus"
